move iconhelper component to injection, update references, fix search icon for pre-v3.0
