--- assets/targets/components/icons/index.scss
@@ -1,1 +1,0 @@
-@import 'svg';
