+++ assets/targets/components/index.scss
@@ -53,6 +53,7 @@ body {
 @import './gallery/index';
 @import './article/index';
 @import './notices/index';
+@import '../injection/icons/svg';
 
 // Print
 @import './print';
