+++ assets/targets/components/media/index.scss
@@ -1,4 +1,3 @@
 @import 'images';
 @import 'yt';
 @import 'soundcloud';
-@import '../icons/svg';
