+++ assets/targets/debranded/index.scss
@@ -52,6 +52,7 @@ body {
 @import '../components/gallery/index';
 @import '../components/article/index';
 @import '../components/notices/index';
+@import '../injection/icons/svg';
 
 // Print
 @import '../components/print';
