+++ assets/targets/injection/icons/_svg.scss
