+++ assets/targets/injection/icons/index.js
