+++ assets/targets/injection/index.scss
@@ -48,6 +48,7 @@ body {
 @import 'footer/index';
 @import 'announcement/index';
 @import 'nav/index';
+@import 'icons/svg';
 
 .hidden {
   display: none;
